15:07

+++ data_structures_docker/Data-Structures/Stack_and_Queue/hello.c
@@ -3,7 +3,7 @@
 #include <limits.h>
 
 int main(void) {
-    uint64_t u = UINT32_MAX; // 0xFFFFFFFF, 4294967295
+    uint64_t u = UINT64_MAX;
     u = u + 1;               // 정의된 래핑 -> 0
     printf("%u\n", u);       // 0
     printf("0x%08x\n", u);   // 0x00000000
